backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyPDF2
passlib
python-jose
PyJWT
joblib
scikit-learn
python-docx
odfpy
